refactor(prodserver): move `setupProxy.js` into prodservers
Including this file in each `prodserver` package means we can keep only the `build` output from each React app and still run the `prodserver` package properly. Symlinks from the old location to keep compatibility. Adds `http-proxy-middleware` to all `prodserver` packages so they can use it since they now own their `setupProxy.js` files.
