+++ apps/bas/package.json
@@ -61,7 +61,7 @@
     "@types/react-dom": "^17.0.0",
     "@types/styled-components": "^5.1.7",
     "@votingworks/utils": "workspace:*",
-    "http-proxy-middleware": "^1.0.0",
+    "http-proxy-middleware": "1.0.6",
     "luxon": "^1.27.0",
     "normalize.css": "^8.0.1",
     "pluralize": "^8.0.0",
