+++ apps/election-manager/package.json
@@ -96,7 +96,7 @@
     "dompurify": "^2.0.12",
     "fetch-mock": "^9.10.7",
     "history": "^4.10.1",
-    "http-proxy-middleware": "^1.0.0",
+    "http-proxy-middleware": "1.0.6",
     "i18next": "^19.4.5",
     "js-file-download": "^0.4.12",
     "js-sha256": "^0.9.0",
