+++ apps/precinct-scanner/package.json
@@ -87,7 +87,7 @@
     "base64-js": "^1.3.1",
     "debug": "^4.3.1",
     "fetch-mock": "^9.11.0",
-    "http-proxy-middleware": "^1.0.0",
+    "http-proxy-middleware": "1.0.6",
     "i18next": "^19.4.5",
     "jest-fetch-mock": "^3.0.3",
     "js-file-download": "^0.4.12",
